// packages/lib/types/apiResponses.ts
export type ApexSuccessResponse = {
  errorcode: 0;
  errormsg: null;
  detail: null;
  result?: boolean;
  data?: unknown;
};

export type ApexErrorResponse = {
  errorcode: number;
  errormsg: string;
  detail?: string | null;
  result?: false;
};

export type ApexResponse = ApexSuccessResponse | ApexErrorResponse;
// ...
export interface AuthResponseType {
  AccountId: string;
  Authenticated: boolean;
  OMSId: string;
  SessionToken: string;
  Token: string;
  UserId: number;
  errormsg?: string;
  Requires2FA?: boolean;
  Pending2FaToken?: string;
}

export interface UserInfo {
  userName: string;
  email: string;
  passwordHash: string;
}

export interface RegisterUserResponseType {
  userInfo: UserInfo;
  userConfig: any[];
  operatorId: number;
}

export type ResetPasswordType = { password: string } | { UserName: string };

export interface ResetPasswordResponseType {
  errormsg?: string;
  errorcode?: number;
  result: boolean;
}

type ProductLevelConfig = {
  VerificationLevel: number;
  ProductId: number;
  ProductName: string; // symbol
  AutoWithdrawThreshold: number;
  DailyDepositLimit: number;
  DailyDepositNotionalLimit: number;
  MonthlyDepositLimit: number;
  MonthlyDepositNotionalLimit: number;
  YearlyDepositLimit: number;
  YearlyDepositNotionalLimit: number;
  DailyWithdrawLimit: number;
  DailyWithdrawNotionalLimit: number;
  MonthlyWithdrawLimit: number;
  MonthlyWithdrawNotionalLimit: number;
  YearlyWithdrawLimit: number;
  YearlyWithdrawNotionalLimit: number;
  NotionalProductId: number;
  OverLimitRejected: boolean;
  WithdrawProcessingDelaySec: number;
  DepositTicketWorkflow: string; // e.g. Internal
  WithdrawTicketWorkflow: string; // e.g. Internal
  RequireWhitelistedAddress: boolean;
  AutoAcceptWhitelistedAddress: boolean;
};

export type VerificationLevelConfigResponse = {
  OMSId: number;
  Level: number;
  Products: Array<ProductLevelConfig>;
};

export type TemplateType = {
  AccountProviderId: number;
  AccountProviderName: string;
  TemplateName: string;
};

export type TemplateTypesResponse = {
  TemplateTypes: TemplateType[];
};

export interface IProduct {
  OMSId: number;
  ProductId: number;
  Product: string; // symbol
  ProductType: 'NationalCurrency' | 'CryptoCurrency' | 'Contract';
  ProductFullName: string;
  DecimalPlaces: number;
  IsDisabled: boolean;
  MarginEnabled: boolean;
  MasterDataUniqueProductSymbol: string | null;
  NoFees: boolean;
  TickSize: number;
}

export interface IAccountPosition {
  AccountId: number;
  Amount: number;
  AvailableBalance: number;
  AvailableBalanceNotional: number;
  DecimalPlaces: number;
  Hold: number;
  DepositEnabled: boolean;
  MarginLend: number;
  MarginLiability: number;
  MarginEnabled: boolean;
  NotionalHoldAmount: number;
  NotionalProductId: number;
  NotionalProductSymbol: string;
  NotionalRate: number;
  NotionalValue: number;
  OMSId: number;
  PendingDeposits: number;
  PendingDepositsNotional: number;
  PendingWithdraws: number;
  ProductFullName: string;
  ProductId: number;
  ProductSymbol: string;
  StartOfDayBalanceNotional: number;
  TotalDayDepositNotional: number;
  TotalDayDeposits: number;
  TotalDayInflowsAndOutflowsNotional: number;
  TotalDayTransferNotional: number;
  TotalDayWithdrawNotional: number;
  TotalDayWithdraws: number;
  TotalMonthDepositNotional: number;
  TotalMonthDeposits: number;
  TotalMonthWithdrawNotional: number;
  TotalMonthWithdraws: number;
  TotalYearDepositNotional: number;
  TotalYearDeposits: number;
  TotalYearWithdrawNotional: number;
  TotalYearWithdraws: number;
  WithdrawEnabled: boolean;
}

export interface IInstrument {
  OMSId: number;
  InstrumentId: number;
  Symbol: string;
  Product1: number;
  Product1Symbol: string;
  Product2: number;
  Product2Symbol: string;
  InstrumentType: string;
  VenueInstrumentId: number;
  VenueId: number;
  SortIndex: number;
  SessionStatus: string;
  PreviousSessionStatus: string;
  SessionStatusDateTime: string;
  SelfTradePrevention: boolean;
  QuantityIncrement: number;
  PriceIncrement: number;
  MinimumQuantity: number;
  MinimumPrice: number;
  VenueSymbol: string;
  IsDisable: boolean;
  MasterDataId: number;
  PriceCollarThreshold: number;
  PriceCollarPercent: number;
  PriceCollarEnabled: boolean;
  PriceFloorLimit: number;
  PriceFloorLimitEnabled: boolean;
  PriceCeilingLimit: number;
  PriceCeilingLimitEnabled: boolean;
  CreateWithMarketRunning: boolean;
  AllowOnlyMarketMakerCounterParty: boolean;
  PriceCollarIndexDifference: number;
  PriceCollarConvertToOtcEnabled: boolean;
  PriceCollarConvertToOtcClientUserId: number;
  PriceCollarConvertToOtcAccountId: number;
  PriceCollarConvertToOtcThreshold: number;
  OtcConvertSizeThreshold: number;
  OtcConvertSizeEnabled: boolean;
  OtcTradesPublic: boolean;
  PriceTier: number;
}

export interface ILevel1 {
  OMSId: number;
  InstrumentId: number;
  BestBid: number;
  BestOffer: number;
  LastTradedPx: number;
  LastTradedQty: number;
  LastTradeTime: number;
  SessionOpen: number;
  SessionHigh: number;
  SessionLow: number;
  SessionClose: number;
  Volume: number;
  CurrentDayVolume: number;
  CurrentDayNotional: number;
  CurrentDayNumTrades: number;
  CurrentDayPxChange: number;
  Rolling24HrVolume: number;
  Rolling24HrNotional: number;
  Rolling24NumTrades: number;
  Rolling24HrPxChange: number;
  TimeStamp: string;
  BidQty: number;
  AskQty: number;
  BidOrderCt: number;
  AskOrderCt: number;
  Rolling24HrPxChangePercent: number;
}

export interface IInstrumentWithLevel1 extends IInstrument, ILevel1 {}

export type OrderSide = 'Buy' | 'Sell' | 'Short' | 'Unknown';
export type OrderType =
  | 'Unknown'
  | 'Market order'
  | 'Limit'
  | 'StopMarket'
  | 'StopLimit'
  | 'TrailingStopMarket'
  | 'TrailingStopLimit'
  | 'BlockTrade';
// ...
export type TradeHistoryItem = {
  OMSId: number;
  ExecutionId: number;
  TradeId: number;
  OrderId: number;
  AccountId: number;
  AccountName: string;
  SubAccountId: number;
  ClientOrderId: number;
  InstrumentId: number;
  Side: OrderSide;
  OrderType: OrderType;
  Quantity: number;
  RemainingQuantity: number;
  Price: number;
  Value: number;
  CounterParty: string;
  OrderTradeRevision: number;
  Direction: string;
  IsBlockTrade: boolean;
  Fee: number;
  FeeProductId: number;
  OrderOriginator: number;
  UserName: string;
  TradeTime: number;
  TradeTimeMS: number;
  MakerTaker: string;
  AdapterTradeId: number;
  InsideBid: number;
  InsideBidSize: number;
  InsideAsk: number;
  InsideAskSize: number;
  IsQuote: boolean;
  CounterPartyClientUserId: number;
  NotionalProductId: number;
  NotionalRate: number;
  NotionalValue: number;
  NotionalHoldAmount: number;
};

export interface IValidateToken {
  data?: {
    UserId: number;
    SessionToken: string;
    Authenticated: boolean;
  };
  error?: string;
}

export interface IUserInfo {
  data?: {
    UserName: string;
    AccountId: number;
  };
}

export type UserInfoResponse = {
  UserId: number;
  UserName: string;
  Email: string;
  EmailVerified: boolean;
  PasswordHash?: string;
  Salt?: string;
  PendingEmailCode?: string;
  PendingCodeTime?: string;
  AccountId: number;
  DateTimeCreated: string;
  AffiliateId?: number;
  RefererId?: number;
  OMSId: number;
  Use2FA: boolean;
  Locked: boolean;
  LockedTime: string;
  NumberOfFailedAttempt: number;
  MarginBorrowerEnabled: boolean;
  MarginAcquisitionHalt: boolean;
  OperatorId: number;
};
// ...
export type SendOrderSuccessResponse = {
  status: 'Accepted' | 'Rejected';
  OrderId: number;
};

export type TransferRequest = {
  CreatedTimestamp: string;
  LastUpdateTimestamp: string;
  OMSId: number;
  OperatorId: number;
  RequestId: number;
  RequestCode: string;
  PayerUsername: string;
  PayerAccountId: number;
  RequestorUsername: string;
  RequestorAccountId: number;
  ProductId: number;
  ProductName: string;
  Amount: number;
  Notes: string;
  Status: string;
};
// ...
export type WithdrawInfoResponse = Array<{
  AccountWithdrawInfoId: string;
  AccountId: number;
  ProductId: number;
  ProductName: string; // it's product full name
  Name: string;
  AccountIdentifier: string;
  Verified: boolean;
}>;
// ...
export interface IMarginProductConfig {
  APY: number;
  CollateralEnabled: boolean;
  HaircutAmount: number;
  HourlyInterestRate: number;
  OMSId: number;
  ProductId: number;
}